fix: Americanize country agnostic code
