+++ src/pages/policy/output/laborSupply/LaborSupplyDecileRelativeImpacts.jsx
@@ -1,20 +1,20 @@
 import { formatPercent } from "../../../../lang/format";
 import {
-  LabourSupplyDecileIncome,
-  LabourSupplyDecileSubstitution,
-  LabourSupplyDecileTotal,
-} from "./LabourSupplyDecileCharts";
+  LaborSupplyDecileIncome,
+  LaborSupplyDecileSubstitution,
+  LaborSupplyDecileTotal,
+} from "./LaborSupplyDecileCharts";
 
-export function LabourSupplyDecileRelativeImpactIncome(props) {
+export function LaborSupplyDecileRelativeImpactIncome(props) {
   const { policyLabel, impact, countryId } = props;
 
-  const decileImpact = impact.labour_supply_response;
+  const decileImpact = impact.labor_supply_response;
 
   const data = decileImpact.decile.relative;
 
   const incomeChanges = Object.values(data.income).slice(0, 10);
 
-  const title = `${policyLabel}'s income effect-driven relative labor supply impact by decile`;
+  const title = `${policyLabel}'s income effect-driven relative ${countryId === "us" ? "labor" : "labour"} supply impact by decile`;
   const description =
     "This chart shows only the income effect-driven portion of " +
     "the estimated relative change in earnings (as a percentage " +
@@ -32,7 +32,7 @@ export function LabourSupplyDecileRelativeImpactIncome(props) {
   };
 
   const chart = (
-    <LabourSupplyDecileIncome
+    <LaborSupplyDecileIncome
       title={title}
       incomeChanges={incomeChanges}
       countryId={countryId}
@@ -46,15 +46,15 @@ export function LabourSupplyDecileRelativeImpactIncome(props) {
   return { chart: chart, csv: () => {} };
 }
 
-export function LabourSupplyDecileRelativeImpactSubstitution(props) {
+export function LaborSupplyDecileRelativeImpactSubstitution(props) {
   const { policyLabel, impact, countryId } = props;
 
-  const decileImpact = impact.labour_supply_response;
+  const decileImpact = impact.labor_supply_response;
 
   const data = decileImpact.decile.relative;
 
   let substitutionChanges = Object.values(data.substitution).slice(0, 10);
-  const title = `${policyLabel}'s substitution effect-driven relative labor supply impact by decile`;
+  const title = `${policyLabel}'s substitution effect-driven relative ${countryId === "us" ? "labor" : "labour"} supply impact by decile`;
   const description =
     "This chart shows only the substitution effect-driven portion of " +
     "the estimated relative change in earnings (as a percentage " +
@@ -71,7 +71,7 @@ export function LabourSupplyDecileRelativeImpactSubstitution(props) {
   };
 
   const chart = (
-    <LabourSupplyDecileSubstitution
+    <LaborSupplyDecileSubstitution
       title={title}
       substitutionChanges={substitutionChanges}
       countryId={countryId}
@@ -85,10 +85,10 @@ export function LabourSupplyDecileRelativeImpactSubstitution(props) {
   return { chart: chart, csv: () => {} };
 }
 
-export function LabourSupplyDecileRelativeImpactTotal(props) {
+export function LaborSupplyDecileRelativeImpactTotal(props) {
   const { policyLabel, impact, countryId } = props;
 
-  const decileImpact = impact.labour_supply_response;
+  const decileImpact = impact.labor_supply_response;
 
   const data = decileImpact.decile.relative;
 
@@ -99,7 +99,7 @@ export function LabourSupplyDecileRelativeImpactTotal(props) {
     overallChange.push(incomeChanges[i] + substitutionChanges[i]);
   }
 
-  const title = `${policyLabel}'s relative labor supply impact by decile`;
+  const title = `${policyLabel}'s relative ${countryId === "us" ? "labor" : "labour"} supply impact by decile`;
   const description =
     "This chart shows the estimated relative change in earnings (as a " +
     "percentage of total earnings) for each disposable income decile.";
@@ -115,7 +115,7 @@ export function LabourSupplyDecileRelativeImpactTotal(props) {
   };
 
   const chart = (
-    <LabourSupplyDecileTotal
+    <LaborSupplyDecileTotal
       title={title}
       overallChange={overallChange}
       countryId={countryId}
