+++ src/pages/policy/output/laborSupply/LaborSupplyHoursImpact.jsx
@@ -1,12 +1,15 @@
-import React from "react";
+import React, { useContext } from "react";
 import Plot from "react-plotly.js";
 import { ChartLogo } from "../../../../api/charts";
-import { formatPercent, localeCode } from "../../../../lang/format";
+import {
+  formatNumberAbbr,
+  formatPercent,
+  localeCode,
+} from "../../../../lang/format";
 import style from "../../../../style";
 import { plotLayoutFont } from "pages/policy/output/utils";
 import ImpactChart, { regionName } from "../ImpactChart";
 import { ChartWidthContext } from "../../../../layout/HoverCard";
-import { useContext } from "react";
 
 function ImpactPlot(props) {
   const { values, labels, metadata, mobile } = props;
@@ -31,9 +34,8 @@ function ImpactPlot(props) {
               : ["total"],
           textposition: "inside",
           text: values.map((value) =>
-            formatPercent(value, metadata.countryId, {
-              maximumFractionDigits: 2,
-              minimumFractionDigits: 2,
+            formatNumberAbbr(value, metadata.countryId, {
+              maximumFractionDigits: 1,
             }),
           ),
           increasing: { marker: { color: style.colors.BLUE } },
@@ -59,8 +61,8 @@ function ImpactPlot(props) {
           title: "",
         },
         yaxis: {
-          title: "Relative change",
-          tickformat: ".1%",
+          title: "Full time-equivalent hours",
+          tickformat: ",.0f",
           fixedrange: true,
         },
         hoverlabel: {
@@ -94,10 +96,9 @@ function ImpactPlot(props) {
 function title(policyLabel, change, metadata) {
   const region = regionName(metadata);
   const regionPhrase = region ? ` in ${region}` : "";
-  const term1 = `earnings${regionPhrase}`;
+  const term1 = `hours worked${regionPhrase}`;
   const term2 = formatPercent(Math.abs(change), metadata.countryId, {
     maximumFractionDigits: 2,
-    minimumFractionDigits: 1,
   });
   const signTerm = change > 0 ? "increase" : "decrease";
   const msg =
@@ -107,19 +108,34 @@ function title(policyLabel, change, metadata) {
   return msg;
 }
 
-export default function lsrImpactRelative(props) {
+export default function lsrHoursImpact(props) {
   const { impact, policyLabel, metadata, mobile } = props;
-
-  const incomeEffect = impact.labour_supply_response.relative_lsr.income;
-  const substitutionEffect =
-    impact.labour_supply_response.relative_lsr.substitution;
-  const netEffect = incomeEffect + substitutionEffect;
-
+  const laborSupply = impact.labor_supply_response;
+  const incomeEffect = laborSupply.hours.income_effect;
+  const substitutionEffect = laborSupply.hours.substitution_effect;
+  const overallEffect = laborSupply.hours.change;
+  const relEffect = overallEffect / laborSupply.hours.baseline;
   const labels = ["Income effect", "Substitution effect", "Net change"];
-  const values = [incomeEffect, substitutionEffect, netEffect];
-
+  const values = [
+    incomeEffect / 40,
+    substitutionEffect / 40,
+    overallEffect / 40,
+  ];
+  const hoursToFTE = (weeklyHours) => weeklyHours / 40;
+  const baselineFTE = hoursToFTE(laborSupply.hours.baseline);
+  const reformFTE = hoursToFTE(laborSupply.hours.reform);
+  const changeFTE = hoursToFTE(overallEffect);
+  const fmt = (x) =>
+    formatNumberAbbr(x, metadata.countryId, {
+      maximumFractionDigits: 1,
+    });
   const chart = (
-    <ImpactChart title={title(policyLabel, netEffect, metadata)}>
+    <ImpactChart title={title(policyLabel, relEffect, metadata)}>
+      <p>
+        This is equivalent to {fmt(changeFTE)},{" "}
+        {overallEffect > 0 ? "increasing" : "decreasing"} full time-equivalent
+        hours from {fmt(baselineFTE)} to {fmt(reformFTE)}.
+      </p>
       <ImpactPlot
         values={values}
         labels={labels}
