+++ src/pages/policy/output/laborSupply/LaborSupplyResponseRelative.jsx
@@ -1,15 +1,12 @@
-import React, { useContext } from "react";
+import React from "react";
 import Plot from "react-plotly.js";
 import { ChartLogo } from "../../../../api/charts";
-import {
-  formatNumberAbbr,
-  formatPercent,
-  localeCode,
-} from "../../../../lang/format";
+import { formatPercent, localeCode } from "../../../../lang/format";
 import style from "../../../../style";
 import { plotLayoutFont } from "pages/policy/output/utils";
 import ImpactChart, { regionName } from "../ImpactChart";
 import { ChartWidthContext } from "../../../../layout/HoverCard";
+import { useContext } from "react";
 
 function ImpactPlot(props) {
   const { values, labels, metadata, mobile } = props;
@@ -34,8 +31,9 @@ function ImpactPlot(props) {
               : ["total"],
           textposition: "inside",
           text: values.map((value) =>
-            formatNumberAbbr(value, metadata.countryId, {
-              maximumFractionDigits: 1,
+            formatPercent(value, metadata.countryId, {
+              maximumFractionDigits: 2,
+              minimumFractionDigits: 2,
             }),
           ),
           increasing: { marker: { color: style.colors.BLUE } },
@@ -61,8 +59,8 @@ function ImpactPlot(props) {
           title: "",
         },
         yaxis: {
-          title: "Full time-equivalent hours",
-          tickformat: ",.0f",
+          title: "Relative change",
+          tickformat: ".1%",
           fixedrange: true,
         },
         hoverlabel: {
@@ -96,9 +94,10 @@ function ImpactPlot(props) {
 function title(policyLabel, change, metadata) {
   const region = regionName(metadata);
   const regionPhrase = region ? ` in ${region}` : "";
-  const term1 = `hours worked${regionPhrase}`;
+  const term1 = `earnings${regionPhrase}`;
   const term2 = formatPercent(Math.abs(change), metadata.countryId, {
     maximumFractionDigits: 2,
+    minimumFractionDigits: 1,
   });
   const signTerm = change > 0 ? "increase" : "decrease";
   const msg =
@@ -108,34 +107,19 @@ function title(policyLabel, change, metadata) {
   return msg;
 }
 
-export default function lsrHoursImpact(props) {
+export default function lsrImpactRelative(props) {
   const { impact, policyLabel, metadata, mobile } = props;
-  const laborSupply = impact.labour_supply_response;
-  const incomeEffect = laborSupply.hours.income_effect;
-  const substitutionEffect = laborSupply.hours.substitution_effect;
-  const overallEffect = laborSupply.hours.change;
-  const relEffect = overallEffect / laborSupply.hours.baseline;
+
+  const incomeEffect = impact.labor_supply_response.relative_lsr.income;
+  const substitutionEffect =
+    impact.labor_supply_response.relative_lsr.substitution;
+  const netEffect = incomeEffect + substitutionEffect;
+
   const labels = ["Income effect", "Substitution effect", "Net change"];
-  const values = [
-    incomeEffect / 40,
-    substitutionEffect / 40,
-    overallEffect / 40,
-  ];
-  const hoursToFTE = (weeklyHours) => weeklyHours / 40;
-  const baselineFTE = hoursToFTE(laborSupply.hours.baseline);
-  const reformFTE = hoursToFTE(laborSupply.hours.reform);
-  const changeFTE = hoursToFTE(overallEffect);
-  const fmt = (x) =>
-    formatNumberAbbr(x, metadata.countryId, {
-      maximumFractionDigits: 1,
-    });
+  const values = [incomeEffect, substitutionEffect, netEffect];
+
   const chart = (
-    <ImpactChart title={title(policyLabel, relEffect, metadata)}>
-      <p>
-        This is equivalent to {fmt(changeFTE)},{" "}
-        {overallEffect > 0 ? "increasing" : "decreasing"} full time-equivalent
-        hours from {fmt(baselineFTE)} to {fmt(reformFTE)}.
-      </p>
+    <ImpactChart title={title(policyLabel, netEffect, metadata)}>
       <ImpactPlot
         values={values}
         labels={labels}
